+++ test/test_battery.rs
@@ -82,8 +82,8 @@ impl EchoServer {
 
         // Register the connection
         self.conns[tok].token = Some(tok);
-        event_loop.register_opt(&self.conns[tok].sock, tok, EventSet::readable(),
-                                PollOpt::edge() | PollOpt::oneshot())
+        event_loop.register(&self.conns[tok].sock, tok, EventSet::readable(),
+                            PollOpt::edge() | PollOpt::oneshot())
             .ok().expect("could not register socket with event loop");
 
         Ok(())
@@ -243,15 +243,15 @@ pub fn test_echo_server() {
     let srv = srv.listen(256).unwrap();
 
     info!("listen for connections");
-    event_loop.register_opt(&srv, SERVER, EventSet::readable(),
-                            PollOpt::edge() | PollOpt::oneshot()).unwrap();
+    event_loop.register(&srv, SERVER, EventSet::readable(),
+                        PollOpt::edge() | PollOpt::oneshot()).unwrap();
 
     let (sock, _) = TcpSocket::v4().unwrap()
         .connect(&addr).unwrap();
 
     // Connect to the server
-    event_loop.register_opt(&sock, CLIENT, EventSet::writable(),
-                            PollOpt::edge() | PollOpt::oneshot()).unwrap();
+    event_loop.register(&sock, CLIENT, EventSet::writable(),
+                        PollOpt::edge() | PollOpt::oneshot()).unwrap();
     let chan = event_loop.channel();
 
     let go = move || {